Edit casting for SAUCE size macros

+++ include/Sauce.h
@@ -170,10 +170,10 @@ enum SAUCE_FileType {
 
 // Determine how large a comment block will be in bytes according to the number of lines present.
 // This includes the 5 bytes for the COMNT id.
-#define SAUCE_COMMENT_BLOCK_SIZE(lines)         ((lines) ? (lines * SAUCE_COMMENT_LINE_LENGTH + 5) : 0)
+#define SAUCE_COMMENT_BLOCK_SIZE(lines)         (((uint8_t)lines) ? ((uint8_t)lines * (uint16_t)SAUCE_COMMENT_LINE_LENGTH + 5) : 0)
 
 // Determine how large a record and optional comment 
-#define SAUCE_TOTAL_SIZE(lines)                 (SAUCE_RECORD_SIZE + SAUCE_COMMENT_BLOCK_SIZE(lines))
+#define SAUCE_TOTAL_SIZE(lines)                 ((uint16_t)SAUCE_RECORD_SIZE + SAUCE_COMMENT_BLOCK_SIZE(lines))
 
 
 // Error Codes
